backend/requirements.txt
fastapi[standard]
torch
pillow
uvicorn[standard]
requests
torchvision
pydantic
requests
geopy